+++ CHANGELOG.md
@@ -5,6 +5,8 @@ FlatLaf Change Log
 
 - ScrollBar: Made styling more flexible by supporting insets and arc for track
   and thumb. (issue #103)
+- ComboBox: Minimum width is now 72 pixels (was ~50 for non-editable and ~130
+  for editable comboboxes).
 - ComboBox: Support custom borders in combobox editors. (issue #102)
 - Ubuntu Linux: Fixed poorly rendered font. (issue #105)